+++ server/src/vscode/Connection.ts
@@ -78,7 +78,7 @@ export class VsCodeConnection implements IConnection {
 
 		this._connection.onInitialize((params: InitializeParams) => {
 			cb({
-				workspaceFolders: params.workspaceFolders,
+				workspaceFolders: params.workspaceFolders ?? null,
 				savePath: params.initializationOptions.savePath,
 				extensionPath: params.initializationOptions.extensionPath,
 				isIDE: true,